Fix dereference passing long instead of integer as index
A ConstantExpression should not be created with value and type mismatch.
The index of a deference call is an integer but will be saved as long
in the ConstantExpression.

+++ presto-main/src/main/java/com/facebook/presto/sql/relational/SqlToRowExpressionTranslator.java
@@ -642,7 +642,7 @@ public final class SqlToRowExpressionTranslator
 
             checkState(index >= 0, "could not find field name: %s", node.getField());
             Type returnType = getType(node);
-            return specialForm(DEREFERENCE, returnType, process(node.getBase(), context), constant(index, INTEGER));
+            return specialForm(DEREFERENCE, returnType, process(node.getBase(), context), constant((long) index, INTEGER));
         }
 
         @Override
